requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bitsandbytes==0.47.0
blinker==1.9.0
boto3==1.40.27
botocore==1.40.27
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
docstring_parser==0.17.0
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
jax==0.7.1
jaxlib==0.7.1
jellyfish==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.5.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
peft==0.11.1
pillow==11.3.0
praat-parselmouth==0.4.6
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
s3transfer==0.14.0
safetensors==0.6.2
scipy==1.16.1
segtok==1.5.11
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.47.3
streamlit==1.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
textstat==0.7.10
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
trl==0.9.4
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yake==0.6.0
yarl==1.20.1
